카드 혜택 상세 조회 API 구현

+++ Backend/fitcard/src/main/java/com/fitcard/domain/card/benefit/model/dto/response/FinancialCardBenefitResponse.java
@@ -43,7 +43,7 @@ public class FinancialCardBenefitResponse {
     private int benefitPer;
 
     @NotNull
-    private int merchantId;
+    private Long merchantId;
 
     @NotEmpty
     private String exceptionTypeList;
